requirements.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Babel==4.0.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
pyodbc==5.0.1
SQLAlchemy==2.0.23

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
WTForms==3.1.1

# File Handling
Pillow==10.1.0

# Translation APIs
google-cloud-translate==3.12.1
requests==2.31.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Logging & Monitoring
python-json-logger==2.0.7
